Add files via upload
Quran Player is a comprehensive website designed to offer a rich and beneficial experience for Quran enthusiasts. The site features a range of outstanding features, including:

Quran Recitations: Enjoy listening to the recitations of the Holy Quran by some of the most renowned reciters from around the world. You can choose your favorite reciter and listen to any Surah you wish.

Quran Tafseer: The site provides a complete interpretation of all Surahs and verses of the Quran, helping users understand the profound meanings and messages of the verses.

Hadith Collection: The site includes a collection of authentic Hadiths that contribute to enriching religious knowledge and guiding users to the teachings of the Prophet Muhammad (PBUH).

Supplications and Remembrances: The site offers a variety of supplications and remembrances that users can benefit from in their daily lives and enhance their spirituality.

Quran Player brings together recitations, interpretations, Hadiths, and supplications in one platform, making it the ideal choice for those interested in the Quran and its sciences. Enjoy a comprehensive and user-friendly religious experience with Quran Player.